Created the Route Page

+++ src/Pages/RoutesPage/RoutesPage.tsx
@@ -12,6 +12,7 @@ import EmptyList from "../../Components/EmptyList/EmptyList";
 import { GoDiffAdded } from "react-icons/go";
 import { CgPlayListRemove } from "react-icons/cg";
 import { deleteRouteById, getAllRoutes } from "../../Services/main.services";
+import { confirmAlert } from "react-confirm-alert";
 
 const RoutesPage = (props: { isLoaded: Boolean }) => {
   const navigate = useNavigate();
@@ -21,19 +22,61 @@ const RoutesPage = (props: { isLoaded: Boolean }) => {
   const [routesToDisplay, setRoutesToDisplay] = useState<BusRoute[]>();
   const [deleteMode, setDeleteMode] = useState<boolean>(false);
 
-  const handleDelete = (id: number) => {
+  const handleDelete = (id: number, label : string) => {
     
     const deleteRoute = async () => {
       await deleteRouteById(id);
     };
 
     if (id) {
-      deleteRoute();
-      setRoutesToDisplay(
-        routesToDisplay?.filter((element) => {
-          return element.route_id !== id;
-        })
-      );
+      confirmAlert({
+        customUI: ({ onClose }) => {
+          return (
+            <div
+              className="custom-ui"
+              style={{
+                display: "flex",
+                flexDirection: "column",
+                alignItems: "center",
+              }}
+            >
+              <h1>Borrar ruta {label}</h1>
+              <h3 style={{ marginBottom: 10, fontWeight: "normal" }}>
+                ¿Quieres borrar esta ruta?
+              </h3>
+              <div
+                style={{
+                  display: "flex",
+                  alignItems: "center",
+                  justifyContent: "space-evenly",
+                }}
+              >
+                <button
+                  style={{ paddingRight: 10, paddingLeft: 10 }}
+                  onClick={() => {
+                    deleteRoute();
+                    setRoutesToDisplay(
+                      routesToDisplay?.filter((element) => {
+                        return element.route_id !== id;
+                      })
+                    );
+                    onClose();
+                  }}
+                >
+                  Confirmar
+                </button>
+                <button
+                  style={{ paddingRight: 10, paddingLeft: 10, marginLeft: 10 }}
+                  onClick={onClose}
+                >
+                  Cancelar
+                </button>
+              </div>
+            </div>
+          );
+        },
+      });
+
     }
   };
